Update 08-webservice-integration.md
Edited copy .

question 1: am i right in thinking that you don't need Matic for gas in this stage? If so, we should remove point 2 in Getting Ready. 

question 2: Is 'using typescript' an alternative to 'using curl'? If so, we should get rid of line 45. And add this to line 42: 'Below are instructions for using curl and for using typescript, but the same works for any HTTP client application or library'.